requirements.txt
streamlit==1.28.0
opencv-python==4.8.1.78
numpy==1.26.4
Pillow==10.4.0
nibabel==5.2.0
